refactor: extract queueStaleAnalysesOnStartup to domain-analysis-startup.ts (file size limit)
domain-analysis-cache.ts exceeded the 700-line hard limit after the startup
warming additions. Moved queueStaleAnalysesOnStartup into its own module and
exported getCurrentSnapshot so the new file can use it.
