+++ cloud/apps/api/src/services/analysis/domain-analysis-cache.ts
@@ -33,11 +33,11 @@ import {
   buildSnapshotOutput,
 } from './domain-analysis-snapshot-builder.js';
 import type { DomainAnalysisScope } from './domain-analysis-scope.js';
-import { DOMAIN_ANALYSIS_ALL_DOMAINS_SCOPE, buildDomainAnalysisDomainSetId, normalizeDomainIds } from './domain-analysis-scope.js';
+import { DOMAIN_ANALYSIS_ALL_DOMAINS_SCOPE } from './domain-analysis-scope.js';
 
 const log = createLogger('analysis:domain-cache');
 
-async function getCurrentSnapshot(
+export async function getCurrentSnapshot(
   client: SnapshotClient,
   scope: DomainAnalysisScope,
   domainId: string,
@@ -631,78 +631,7 @@ export async function getDomainAnalysisResult(params: {
   });
 }
 
-/**
- * On startup, queue background refreshes for priority analysis combinations that are
- * stale or have never been built. Priority set:
- *   1. ALL_DOMAINS
- *   2. Every individual domain
- *   3. All domains minus "Motivation for Invasion Choice" (the most-used domain set)
- *
- * Runs non-blocking after the queue orchestrator is ready.
- * Skips any combination whose rebuild is already in progress.
- */
-export async function queueStaleAnalysesOnStartup(): Promise<void> {
-  const domains = await db.domain.findMany({
-    select: { id: true, name: true },
-  });
 
-  let queued = 0;
-
-  // Helper: check one combination and queue a refresh if stale or never built.
-  async function checkAndQueue(
-    scope: DomainAnalysisScope,
-    domainId: string,
-    domainIds: string[] | undefined,
-    logContext: Record<string, unknown>,
-  ): Promise<void> {
-    try {
-      const state = await prepareDomainAnalysisState({
-        scope,
-        domainId,
-        domainIds,
-        requestedSignature: null,
-      });
-
-      if (state.definitions.length === 0) return;
-
-      const currentSnapshot = await getCurrentSnapshot(db, state.scope, domainId, state.configSignature);
-      const parsedCurrent = currentSnapshot != null ? parseSnapshotOutput(currentSnapshot.output) : null;
-
-      if (parsedCurrent == null && currentSnapshot != null) return; // rebuild already in progress
-
-      if (currentSnapshot != null && currentSnapshot.inputHash === state.inputHash) return; // already fresh
-
-      const didQueue = await queueDomainAnalysisRefresh({
-        scope: state.scope,
-        domainId,
-        domainIds: scope === 'ALL_DOMAINS' ? undefined : domainIds,
-        signature: state.selectedSignature,
-        reason: currentSnapshot == null ? 'startup-never-built' : 'startup-stale',
-      });
-      if (didQueue) queued += 1;
-    } catch (err) {
-      log.warn({ err, ...logContext }, 'startup stale check failed');
-    }
-  }
-
-  // 1. ALL_DOMAINS
-  await checkAndQueue('ALL_DOMAINS', DOMAIN_ANALYSIS_ALL_DOMAINS_SCOPE, undefined, { scope: 'ALL_DOMAINS' });
-
-  // 2. Individual domains
-  for (const domain of domains) {
-    await checkAndQueue('DOMAIN', domain.id, undefined, { scope: 'DOMAIN', domainId: domain.id });
-  }
-
-  // 3. All domains minus "Motivation for Invasion Choice" (most-used multi-domain combination)
-  const motivationDomain = domains.find((d) => d.name.includes('Motivation for Invasion'));
-  if (motivationDomain != null) {
-    const domainSetIds = normalizeDomainIds(domains.filter((d) => d.id !== motivationDomain.id).map((d) => d.id));
-    const domainSetId = buildDomainAnalysisDomainSetId(domainSetIds);
-    await checkAndQueue('DOMAIN_SET', domainSetId, domainSetIds, { scope: 'DOMAIN_SET', size: domainSetIds.length });
-  }
-
-  log.info({ queued }, 'Startup priority domain analysis check complete');
-}
 
 /**
  * Read the pre-computed per-(definitionId::modelId::canonicalA::canonicalB::ownLevel::opponentLevel)
